+++ engine/include/ui/button/button.h
@@ -16,7 +16,7 @@ class Button: public Object2D
 public:
     Button() = default;
     Button(std::string text, bool is_togglable = false,
-        Vector2<int> position = {}, Vector2<int> size = {64},
+        Vector2<float> position = {}, Vector2<int> size = {64},
         TTF_Font* font = Engine::default_font);
     
     virtual ~Button() override;
@@ -24,7 +24,7 @@ public:
     void render(SDL_Renderer* m_renderer) override;
     void handle_event(const SDL_Event& event) override;
 
-    void set_position(const Vector2<int>& position) override;
+    void set_position(const Vector2<float>& position) override;
 
     const bool& get_is_pressed() const;
     void set_is_pressed(const bool& pressed);
